refactor: DTO files and annotation refactoring
- DTO 파일들을 PortoflioResponse 파일로 통합
- 어노테이션 정리
