--- sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/dto/response/PortfolioListItemDTO.java
@@ -1,12 +1,0 @@
-package com.kakao.sunsuwedding.portfolio.dto.response;
-
-public record PortfolioListItemDTO(
-        Long id,
-        String image,
-        String title,
-        String plannerName,
-        Long price,
-        String location,
-        Long contractCount
-) {
-}
