--- sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/dto/response/PriceDTO.java
@@ -1,9 +1,0 @@
-package com.kakao.sunsuwedding.portfolio.dto.response;
-
-import java.util.List;
-
-public record PriceDTO(
-        Long totalPrice,
-        List<PriceItemDTO> itemPrices
-) {
-}
