--- sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/dto/response/PriceItemDTO.java
@@ -1,7 +1,0 @@
-package com.kakao.sunsuwedding.portfolio.dto.response;
-
-public record PriceItemDTO(
-        String itemTitle,
-        Long itemPrice
-) {
-}
