+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/image/ImageEncoder.java
@@ -2,6 +2,7 @@ package com.kakao.sunsuwedding.portfolio.image;
 
 import com.kakao.sunsuwedding._core.errors.BaseException;
 import com.kakao.sunsuwedding._core.errors.exception.Exception404;
+import com.kakao.sunsuwedding._core.errors.exception.Exception500;
 import org.springframework.core.io.FileSystemResource;
 import org.springframework.core.io.Resource;
 
@@ -19,9 +20,7 @@ public class ImageEncoder {
             return Base64.getEncoder().encodeToString(resource.getContentAsByteArray());
         }
         catch (IOException exception) {
-            exception.printStackTrace();
+            throw new Exception500("이미지 인코딩 과정에서 오류가 발생했습니다.");
         }
-
-        return null;
     }
 }
